refactor: flatten `UserSettings` component file structure
