+++ packages/console/src/pages/UserDetails/UserSettings/PersonalAccessTokens/CreateTokenModal.tsx
