+++ packages/console/src/pages/UserDetails/UserSettings/PersonalAccessTokens/index.module.scss
